+++ CHANGES.rst
@@ -4,8 +4,8 @@ Changelog
 0.3 (unreleased)
 ----------------
 
-- TBD
-
+- Drop the trailing padding (``=``) from base64 content.
+  See https://github.com/Pylons/pyramid_nacl_session/pull/7
 
 0.2 (2015-11-23)
 ----------------