+++ CHANGES.rst
@@ -9,6 +9,10 @@ Changelog
 - Drop the trailing padding (``=``) from base64 content.
   See https://github.com/Pylons/pyramid_nacl_session/pull/7
 
+- Add the ``EncryptedCookieSessionFactory`` as the primary API for using
+  an encrypted session.
+  See https://github.com/Pylons/pyramid_nacl_session/pull/6
+
 0.2 (2015-11-23)
 ----------------